Refactor SpeakCommands to another file
